+++ src/Sentry.Unity/SentryUnity.cs
@@ -32,7 +32,10 @@ namespace Sentry.Unity
             // Uses the game `version` as Release unless the user defined one via the Options
             if (unitySentryOptions.Release == null)
             {
-                unitySentryOptions.Release = $"{Application.productName}@{Application.version}";
+                unitySentryOptions.Release = Application.productName is string productName
+                    && !string.IsNullOrWhiteSpace(productName)
+                        ? $"{productName}@{Application.version}"
+                        : $"{Application.version}";
 
                 unitySentryOptions.DiagnosticLogger?.Log(SentryLevel.Debug,
                     "Setting Sentry Release to Unity App.Version: {0}",