+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixes
 
+- Fallback for missing/empty product name ([#205](https://github.com/getsentry/sentry-unity/pull/205))
 - Add product name to release as default ([#202](https://github.com/getsentry/sentry-unity/pull/202))
 - normalize line endings ([#204](https://github.com/getsentry/sentry-unity/pull/204))